pyproject.toml
[project]
name = "nacos-export"
version = "1.5.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]
